check if files need removeal

+++ custom/bin/backup_home.sh
@@ -88,6 +88,8 @@ if [ "$1" = "-l" ];then
 	if [ ! -d $ENCRYPT_DIR ];then mkdir -p $ENCRYPT_DIR;fi
 	pushd $ENCRYPT_DIR > /dev/null
 	encrypt_ssh.sh $backup_file
-	ls -t $ENCRYPT_DIR/${BACKUP_NAME}_*.tgz.enc |tail -n +2 | xargs rm --
-	ls -t $ENCRYPT_DIR/${BACKUP_NAME}_*.tgz.key.enc |tail -n +2 | xargs rm --
+	if ls $ENCRYPT_DIR/${BACKUP_NAME}_*.tgz.enc 1>/dev/null 2>&1;then
+		ls -t $ENCRYPT_DIR/${BACKUP_NAME}_*.tgz.enc |tail -n +2 | xargs rm --
+		ls -t $ENCRYPT_DIR/${BACKUP_NAME}_*.tgz.key.enc |tail -n +2 | xargs rm --
+	fi
 fi
